# task-management/generated/src/main/java/org/openapitools/api/ApiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.LoginRequestDTO;
import org.openapitools.model.TaskRequestDTO;
import org.openapitools.model.TaskResponseDTO;
import org.openapitools.model.TokenResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-15T21:03:15.372313934Z[Etc/UTC]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "task-controller", description = "the task-controller API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PUT /api/tasks/{id}/complete
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "completeTask",
        tags = { "task-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = TaskResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/tasks/{id}/complete",
        produces = { "*/*" }
    )
    
    default ResponseEntity<TaskResponseDTO> completeTask(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/tasks
     *
     * @param taskRequestDTO  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createTask",
        tags = { "task-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = TaskResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/tasks",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<TaskResponseDTO> createTask(
        @Parameter(name = "TaskRequestDTO", description = "", required = true) @Valid @RequestBody TaskRequestDTO taskRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/tasks/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteTask",
        tags = { "task-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = TaskResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/tasks/{id}",
        produces = { "*/*" }
    )
    
    default ResponseEntity<TaskResponseDTO> deleteTask(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/tasks
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getAllTasks",
        tags = { "task-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = TaskResponseDTO.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/tasks",
        produces = { "*/*" }
    )
    
    default ResponseEntity<List<TaskResponseDTO>> getAllTasks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "[ { \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 }, { \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 } ]";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/tasks/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getTaskById",
        tags = { "task-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = TaskResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/tasks/{id}",
        produces = { "*/*" }
    )
    
    default ResponseEntity<TaskResponseDTO> getTaskById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/auth/login
     *
     * @param loginRequestDTO  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "loginUser",
        tags = { "auth-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = TokenResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/login",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<TokenResponseDTO> loginUser(
        @Parameter(name = "LoginRequestDTO", description = "", required = true) @Valid @RequestBody LoginRequestDTO loginRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"token\" : \"token\" }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/auth/register
     *
     * @param loginRequestDTO  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "registerUser",
        tags = { "auth-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = Object.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/register",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Object> registerUser(
        @Parameter(name = "LoginRequestDTO", description = "", required = true) @Valid @RequestBody LoginRequestDTO loginRequestDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /api/tasks/{id}
     *
     * @param id  (required)
     * @param taskRequestDTO  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateTask",
        tags = { "task-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = TaskResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/tasks/{id}",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<TaskResponseDTO> updateTask(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "TaskRequestDTO", description = "", required = true) @Valid @RequestBody TaskRequestDTO taskRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : 0, \"completed\" : true, \"title\" : \"title\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}